backend/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pillow>=10.0.0,<11.0.0
pydantic>=2.0.0,<3.0.0
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.24.0,<1.0.0